First

+++ fcm.cpp
@@ -4,7 +4,7 @@
 int main() {
     ofstream otwts("test-data/weights.csv");
 	//fitness();
-    array<double, 100> temp = pso(0.2,0.5,5.0,5.0);
+    array<long double, 100> temp = pso(0.2,0.5,5.0,5.0);
     for(int m = 0; m < 10; m++) { //import the weights
         for(int n = 0; n < 10; n++) {
             otwts << temp[m*10 + n] << "\t";
